+++ SuperMarioBros3/RaccoonMario.cpp
@@ -74,8 +74,8 @@ void RaccoonMario::Update(DWORD dt, vector<LPGAMEOBJECT>* coObjects)
 
 void RaccoonMario::OnNoCollision(DWORD dt)
 {
-	mario->position.x += mario->GetSpeed().x * dt;
-	mario->position.y += mario->GetSpeed().y * dt;
+	mario->position.x += (float)mario->GetSpeed().x * dt;
+	mario->position.y += (float)mario->GetSpeed().y * dt;
 }
 
 void RaccoonMario::OnCollisionWith(LPCOLLISIONEVENT e)
