+++ SuperMarioBros3/SmallMario.cpp
@@ -47,8 +47,8 @@ void SmallMario::Update(DWORD dt, vector<LPGAMEOBJECT>* coObjects)
 
 void SmallMario::OnNoCollision(DWORD dt)
 {
-	mario->position.x += mario->GetSpeed().x * dt;
-	mario->position.y += mario->GetSpeed().y * dt;
+	mario->position.x += (float)mario->GetSpeed().x * dt;
+	mario->position.y += (float)mario->GetSpeed().y * dt;
 }
 
 void SmallMario::OnCollisionWith(LPCOLLISIONEVENT e)
